requirements.txt
ultralytics
stremlit
Pillow
numpy